% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_type.R, R/format_type.R
\name{analyze_type}
\alias{analyze_type}
\title{Analyze the Type of a Column}
\usage{
analyze_type(data, col_name, data_type)

analyze_type(data, col_name, data_type)
}
\arguments{
\item{data}{The name of the dataset}

\item{col_name}{The name of the column in the dataset, expressed as a string}

\item{data_type}{The data type you are looking for. This should be one of the four strings: "integer", "numeric", "boolean", or "character"}
}
\value{
A table displaying all of the columns that are unable to be coerced into the chosen type

A table displaying all of the columns that are unable to be coerced into the chosen type
}
\description{
Takes a column from a dataframe and returns a table that shows any rows that are not able to be coerced into the specified type.
This includes any entries that are NA.

Takes a column from a dataframe and returns a table that shows any rows that are not able to be coerced into the specified type.
This includes any entries that are NA.
}
\examples{
analyze_type(mouse_data_bw_dirty, "Body Weight 3", "numeric")

analyze_type(mouse_data_bw_dirty, "Body Weight 3", "numeric")

}
