% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_data.R
\docType{data}
\name{food_data}
\alias{food_data}
\title{Food Data}
\format{
A data frame with 125 rows and 3 variables:
\describe{
\item{\code{comfort_food}}{Character, list of comfort foods}
\item{\code{food_childhood}}{Character, list of childhood foods}
\item{\code{meals_dinner_friend}}{Character, list of foods eaten when friends came over}
}
}
\source{
https://www.kaggle.com/datasets/borapajo/food-choices/versions/1/data
}
\usage{
food_data
}
\description{
This dataset contains information regarding food from students of Mercyhurst University
}
\keyword{datasets}
