% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_data_birth.R
\docType{data}
\name{mouse_data_birth}
\alias{mouse_data_birth}
\title{Mouse Birth Data}
\format{
A data frame with 32 rows and 4 variables:
\describe{
\item{\code{ID}}{Chatacter, description of the mouse sex and number}
\item{\code{Sex}}{Character, "M" for male, "F" for female}
\item{\code{Num}}{Numeric, unique ID number of the mouse}
\item{\code{Treatment}}{Character, either "Plac" for placebo or "Tmt" for treatment}
}
}
\source{
James Xenakis's Mouse Data
}
\usage{
mouse_data_birth
}
\description{
This dataset contains information on mice at the time of their birth
}
\keyword{datasets}
