% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_data_bw_clean.R
\docType{data}
\name{mouse_data_bw_clean}
\alias{mouse_data_bw_clean}
\title{Mouse Body Weight Data (Clean)}
\format{
A data frame with 32 rows and 7 variables:
\describe{
\item{\code{ID}}{Chatracter, description of the mouse sex and number}
\item{\code{Body Weight 1}}{Numeric, the Body Weight of the mouse at Date Body Weight 1}
\item{\code{Date Body Weight 1}}{Numeric, the date Body Weight 1 was taken in POSIXlt format}
\item{\code{Body Weight 2}}{Numeric, the Body Weight of the mouse at Date Body Weight 2}
\item{\code{Date Body Weight 2}}{Numeric, the date Body Weight 2 was taken in POSIXlt format}
\item{\code{Body Weight 3}}{Numeric, the Body Weight of the mouse at Date Body Weight 3}
\item{\code{Date Body Weight 3}}{Numeric, the date Body Weight 3 was taken in POSIXlt format}
}
}
\source{
James Xenakis's Mouse Data
}
\usage{
mouse_data_bw_clean
}
\description{
This dataset contains cleaned information for the weight of each mouse taken over three dates.
}
\keyword{datasets}
