% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot Outliers}
\usage{
plot_outliers(data, col_names, threshold = 0.2, ignore_NA = FALSE)
}
\arguments{
\item{data}{The name of the dataset}

\item{col_names}{A vector of column names that}

\item{threshold}{A number representing the percent change to classify as an "outlier" (default .2, or 20\% change)}

\item{ignore_NA}{A boolean determining whether to ignore NA values}
}
\value{
A plot displaying each outlier plotted on a graph
}
\description{
This function takes a longitunal dataset and finds any unusually large percent changes between datapoints in each row.
}
\examples{
plot_outliers(mouse_data_bw_clean, (c("Body Weight 1", "Body Weight 2", "Body Weight 3")), .2)

}
