% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_list.R
\name{separate_list}
\alias{separate_list}
\title{Separate List}
\usage{
separate_list(
  data,
  col_name,
  min_frequency = 1,
  delimiter = c(", ", ", and", ",", " and ")
)
}
\arguments{
\item{data}{The name of the dataset}

\item{col_name}{The name of the column, expressed as a string}

\item{min_frequency}{The minimum number of times a string needs to occur before it appears in the table (Default is 1)}

\item{delimiter}{The various words or punctuation marks that are used to separate strings, represented as a vector (Default is c(", ", ", and",",", " and "))}
}
\value{
A table representing the most common strings in the responses
}
\description{
This function takes a column of a dataset which has multiple responses in each row and separates the responses into a table
}
\examples{
separate_list(food_data, "comfort_food", 5)

}
