% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_different.R
\name{table_different}
\alias{table_different}
\title{Analyze String Typos
Creates a table with rows that have different values than the rest of the other rows.}
\usage{
table_different(data, col_name, frequency = 5)
}
\arguments{
\item{data}{dataframe The dataframe of information}

\item{col_name}{character The name of column}

\item{frequency}{numeric The maximum number of times a string can appear before it no longer appears on the table}
}
\value{
Returns a table with a list of row numbers and strings that showed up a fewer number of times than frequency
}
\description{
Analyze String Typos
Creates a table with rows that have different values than the rest of the other rows.
}
\examples{
table_different(mouse_data_birth, "Treatment", 5)

}
