% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remice-package.R
\docType{package}
\name{remice-package}
\alias{remice}
\alias{remice-package}
\title{remice: Tools for Cleaning Mice Data}
\description{
The remice package provides tools for cleaning and analyzing datasets,
including those derived from experiments involving mice.
}
\details{
The documentation for this package is available in the package vignette.
Run \code{vignette("intro", package = "remice")} to see the vignette.
}
\seealso{
\code{\link{vignette}}
}
\author{
\strong{Maintainer}: Eric Tong \email{erictong@college.harvard.edu}

}
\keyword{internal}
